% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{base_packages}
\alias{base_packages}
\title{FUNCTION_TITLE}
\usage{
base_packages()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{installed.packages}}
}
