% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_fns.R
\name{desc2pkgr}
\alias{desc2pkgr}
\title{FUNCTION_TITLE}
\usage{
desc2pkgr(desc_path = ".", src_path = "pkg",
  lib_path = .libPaths()[1], pkgr_tmpl = system.file("DESC.tmpl",
  package = "pkgr.utils"), out = "")
}
\arguments{
\item{desc_path}{PARAM_DESCRIPTION, Default: '.'}

\item{src_path}{PARAM_DESCRIPTION, Default: 'pkg'}

\item{lib_path}{PARAM_DESCRIPTION, Default: .libPaths()[1]}

\item{pkgr_tmpl}{PARAM_DESCRIPTION, Default: system.file("DESC.tmpl", package = "pkgr.utils")}

\item{out}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}
 \code{\link[utils]{download.file}}
 \code{\link[yaml]{as.yaml}}
 \code{\link[whisker]{whisker.render}}
 \code{\link[desc]{desc_get_field}}
}
