% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_fns.R
\name{desc2vec}
\alias{desc2vec}
\title{FUNCTION_TITLE}
\usage{
desc2vec(file = ".")
}
\arguments{
\item{file}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[desc]{desc_get_deps}}
}
