% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_deps}
\alias{get_deps}
\title{Extract DESCRIPTION file dependencies}
\usage{
get_deps(file = ".", type = c("DESCRIPTION", "pkgSetup"), ...)
}
\arguments{
\item{file}{PARAM_DESCRIPTION, Default: '.'}

\item{type}{character, type of file to parse, Default: c('DESCRIPTION','pkgSetup')}

\item{\dots}{arguments to pass to desc_deps or pkgSetup_deps}
}
\value{
character
}
\description{
Extract from a DESCRIPTION file a vector of the listed package
dependencies.
}
\details{
For the vector of base and recommended packages use
\link[pkgr.utils:base_packages]{base_packages}
}
