% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_repos}
\alias{get_repos}
\title{Stored Repositories}
\usage{
get_repos(
  name = c("MPN", "CRAN"),
  snapshot = c(MPN = mpn_release(latest = TRUE), CRAN = Sys.Date() - 1)
)
}
\arguments{
\item{name}{character, Names of repositories to return,
Default: c("MPN", "CRAN")}

\item{snapshot}{character, names vector corresponding to the repository of the dates to use for snapshots,
Default: c(MPN = mpn_release(latest = FALSE), CRAN = Sys.Date() - 1)}
}
\value{
character
}
\description{
Returns names of stored CRAN and CRANlike repositories
}
\details{
Snapshot value is coerced into a Date class internally.

The earliest date for 'MPN' is '2019-09-22', if snapshot is earlier than this and
the user includes 'MPN' in name then snapshot will be replaced with '2019-09-22'.
}
\examples{

get_repos()

get_repos(name = 'MPN')

get_repos(name = 'CRAN', snapshot = c(CRAN = Sys.Date() - 1))


}
