% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{mpn_release}
\alias{mpn_release}
\title{Query MPN releases}
\usage{
mpn_release(latest = FALSE, pat = Sys.getenv("GHE_PAT"))
}
\arguments{
\item{latest}{logical, return only the lasest release, Default: FALSE}

\item{pat}{character, GHE Personal Access Token, Default: Sys.getenv("GHE_PAT")}
}
\value{
if latest is TRUE then a character, else a list
}
\description{
Query GHE for list of MPN releases
}
