% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkgr.R
\name{pkgr.current_release}
\alias{pkgr.current_release}
\title{FUNCTION_TITLE}
\usage{
pkgr.current_release(os = "linux")
}
\arguments{
\item{os}{PARAM_DESCRIPTION, Default: 'linux'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
