% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgr.diff}
\alias{pkgr.diff}
\title{FUNCTION_TITLE}
\usage{
pkgr.diff(pkgsetup = "pkgSetup.R", pkgr = "pkgr.yml")
}
\arguments{
\item{pkgsetup}{PARAM_DESCRIPTION, Default: 'pkgSetup.R'}

\item{pkgr}{PARAM_DESCRIPTION, Default: 'pkgr.yml'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[yaml]{read_yaml}}
}
