% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_commands.R
\name{pkgr.help}
\alias{pkgr.help}
\title{Help for pkgr commands}
\usage{
pkgr.help(cmd)
}
\arguments{
\item{cmd}{character, command}
}
\value{
output to console
}
\description{
Get help documentation for pkgr commands
}
\examples{
\dontrun{
pkgr.help('add')
}
}
