% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgr.here}
\alias{pkgr.here}
\title{Location of pkgr.yml}
\usage{
pkgr.here(path = "pkgr.yml")
}
\arguments{
\item{path}{here path to pkgr.yml, Default: 'pkgr.yml'}
}
\value{
character
}
\description{
Function that uses \code{\link[here]{here}} to build path to pkgr.yml
}
\examples{
pkgr.here()
}
\seealso{
\code{\link[here]{here}}
}
