% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgr.new}
\alias{pkgr.new}
\title{FUNCTION_TITLE}
\usage{
pkgr.new(
  pkgs = NULL,
  libpath = .libPaths()[1],
  version = 1,
  threads = parallel::detectCores() - 1,
  repos = get_repos(),
  pkgr_tmpl = system.file("pkgr.tmpl", package = "pkgr.utils"),
  out = ""
)
}
\arguments{
\item{pkgs}{PARAM_DESCRIPTION, Default: NULL}

\item{libpath}{PARAM_DESCRIPTION, Default: 'lib'}

\item{version}{PARAM_DESCRIPTION, Default: 1}

\item{threads}{PARAM_DESCRIPTION, Default: parallel::detectCores() - 1}

\item{repos}{PARAM_DESCRIPTION, Default: get_repos()}

\item{pkgr_tmpl}{PARAM_DESCRIPTION, Default: system.file("pkgr.tmpl", package = "pkgr.utils")}

\item{out}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
pkgr.new('dplyr')
}
