% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgr.open}
\alias{pkgr.open}
\title{Open pkgr.yml in editor}
\usage{
pkgr.open(path = pkgr.here())
}
\arguments{
\item{path}{character, path to pkgr.yml, Default: pkgr.here()}
}
\description{
Opens a pkgr.file in the RStudio Editor
}
\examples{
\dontrun{
pkgr.open()
}
}
