% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_commands.R
\name{pkgr.version}
\alias{pkgr.version}
\title{pkgr version}
\usage{
pkgr.version()
}
\value{
character
}
\description{
Returns string of installed pkgr cli version
}
\examples{
\dontrun{
pkgr.version()
}
}
