% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSetup_fns.R
\name{pkgSetup2pkgr}
\alias{pkgSetup2pkgr}
\title{FUNCTION_TITLE}
\usage{
pkgSetup2pkgr(pkgsetup = "pkgSetup.R", repos = c(pkgsetup = "pkg",
  mrg_repos, cran_repos), libpath = "lib", ...)
}
\arguments{
\item{pkgsetup}{PARAM_DESCRIPTION, Default: 'pkgSetup.R'}

\item{repos}{PARAM_DESCRIPTION, Default: c(pkgsetup = 'pkg',mrg_repos,cran_repos)}

\item{libpath}{PARAM_DESCRIPTION, Default: 'lib'}

\item{\dots}{arguments to pass to \link[pkgr.utils:pkgr.new]{pkgr.new}}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
pkgSetup <- system.file('pkgSetup.R',package = 'pkgr.utils')
pkgSetup2pkgr(pkgSetup)
}
\seealso{
\code{\link[yaml]{as.yaml}}
\code{\link[whisker]{whisker.render}}
}
