% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_fns.R
\name{update_pkgr_desc}
\alias{update_pkgr_desc}
\title{FUNCTION_TITLE}
\usage{
update_pkgr_desc(desc_file = ".", pkgr_yml, test = TRUE)
}
\arguments{
\item{desc_file}{PARAM_DESCRIPTION, Default: '.'}

\item{pkgr_yml}{PARAM_DESCRIPTION}

\item{test}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[yaml]{read_yaml}},\code{\link[yaml]{as.yaml}},\code{\link[yaml]{write_yaml}}
}
