% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_fns.R
\name{desc2pkgr}
\alias{desc2pkgr}
\title{Convert a Description file to pkgr.yml}
\usage{
desc2pkgr(desc_path = ".", src_path = "pkg",
  lib_path = .libPaths()[1], pkgr_tmpl = system.file("DESC.tmpl",
  package = "pkgr.utils"), out = "")
}
\arguments{
\item{desc_path}{character, path to DESCRIPTION file, Default: '.'}

\item{src_path}{character, path to store source files, Default: 'pkg'}

\item{lib_path}{characer, path to install libraries, Default: .libPaths()[1]}

\item{pkgr_tmpl}{character, path to pkgr template,
Default: system.file("DESC.tmpl", package = "pkgr.utils")}

\item{out}{characer, path to save pkgr template, if '' then it is written
to console, Default: ''}
}
\value{
populated pkgr template
}
\description{
Reads in a DESCRIPTION file and populates a pkgr.yml template
with the non base packages listed as Depends, Imports, Suggests
}
\details{
The path to the DESCRIPTION file can be a web URI, in this case it will
be downloaded and read.
}
\examples{
\dontrun{
 desc2pkgr()
}
}
