% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_fns.R
\name{desc2pkgr}
\alias{desc2pkgr}
\title{Convert a Description file to pkgr.yml}
\usage{
desc2pkgr(desc_path = ".", ...)
}
\arguments{
\item{desc_path}{character, path to DESCRIPTION file, Default: '.'}

\item{\dots}{arguments passed on to \link[pkgr.utils:pkgr.new]{pkgr.new}}
}
\value{
populated pkgr template
}
\description{
Reads in a DESCRIPTION file and populates a pkgr.yml template
with the non base packages listed as Depends, Imports, Suggests
}
\details{
The path to the DESCRIPTION file can be a web URI, in this case it will
be downloaded and read.
}
\examples{
\dontrun{
 desc2pkgr()
}
}
