% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_deps}
\alias{get_deps}
\title{Extract DESCRIPTION file dependencies}
\usage{
get_deps(file = ".", type = c("DESCRIPTION", "pkgSetup"), ...)
}
\arguments{
\item{file}{PARAM_DESCRIPTION, Default: '.'}

\item{excl_base}{logical, exclude the base and recommended packages shipped
with R installation}
}
\value{
character
}
\description{
Extract from a DESCRIPTION file a vector of the listed package
dependencies.
}
\details{
For the vector of base and recommended packages use
\link[pkgr.utils:base_packages]{base_packages}
}
\examples{
desc_deps()
}
\seealso{
\link[desc:desc_get_deps]{desc_get_deps}
}
