% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_repos}
\alias{get_repos}
\title{Stored Repositories}
\usage{
get_repos(name = c("gh_external", "gh_dev", "mrg_val", "rs_cran"))
}
\arguments{
\item{name}{character, Names of repositories to return,
Default: c("gh_external", "gh_dev", "mrg_val", "rs_cran")}
}
\value{
character
}
\description{
Returns names of stored CRAN and CRANlike repositories
}
\examples{
get_repos()
}
