% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgr.diff}
\alias{pkgr.diff}
\title{Compare packages listed in DESCRIPTION/pkgSetup.R and pkgr.yml}
\usage{
pkgr.diff(src = "DESCRIPTION", pkgr = "pkgr.yml")
}
\arguments{
\item{src}{character, path to DESCRIPTION/pkgSetup.R , Default: 'DESCRIPTION'}

\item{pkgr}{character, path to pkgr.yml, Default: 'pkgr.yml'}
}
\value{
character
}
\description{
Compare packages listed in DESCRIPTION/pkgSetup.R and pkgr.yml
and return the setdiff.
}
\details{
The setdiff returns what packages are in DESCRIPTION/pkgSetup.R
that are missing from pkgr.yml Package field.
}
\examples{
\dontrun{

pkgr.diff()

pkgr.diff()\%>\%
  pkgr.add()

}
}
\seealso{
\code{\link[yaml]{read_yaml}}
}
