% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_commands.R
\name{pkgr.add}
\alias{pkgr.add}
\alias{pkgr.remove}
\alias{pkgr.install}
\alias{pkgr.plan}
\alias{pkgr.inspect}
\alias{pkgr.run}
\alias{pkgr.clean}
\title{pkgr commands}
\usage{
pkgr.add(pkgs, ..., flags = NULL)

pkgr.remove(pkgs, ..., flags = NULL)

pkgr.install(..., flags = NULL)

pkgr.plan(..., flags = NULL)

pkgr.inspect(..., flags = NULL)

pkgr.run(..., flags = NULL)

pkgr.clean(..., flags = NULL)
}
\arguments{
\item{pkgs}{character, names of packages passed in add/remove}

\item{\dots}{options that are passed into the command calls}

\item{flags}{character, names of flags that are passed into the command calls}
}
\value{
character
}
\description{
Commands that can be called in pkgr command line
}
\details{
\if{html}{
\out{
<br>
<details>
<summary> <span title='Click to Expand'> pkgr add [package name1] [package name2] ... [options] [flags] </span> </summary>
}
}

add package/s to the configuration file and optionally install

\tabular{ll}{
\strong{Flag} \tab \strong{Description}\cr
install   \tab install package/s after adding
}

\if{html}{
\out{
</details>
}
}

\if{html}{
\out{
<br>
<details>
<summary> <span title='Click to Expand'> pkgr remove [package name1] [package name2] ... [options] [flags] </span> </summary>
}
}

remove package/s from the configuration file

\if{html}{
\out{
</details>
}
}

\if{html}{
\out{
<br>
<details>
<summary> <span title='Click to Expand'> pkgr install [options] [flags] </span> </summary>
}
}

install a package

\if{html}{
\out{
</details>
}
}
\if{html}{
\out{
</details>
}
}

\if{html}{
\out{
<br>
<details>
<summary> <span title='Click to Expand'> pkgr plan [options] [flags] </span> </summary>
}
}

see the plan for an install

\if{html}{
\out{
</details>
}
}

\if{html}{
\out{
</details>
}
}

\if{html}{
\out{
<br>
<details>
<summary> <span title='Click to Expand'> Global options </span> </summary>
}
}

\tabular{llll}{
\strong{Option} \tab \strong{Class} \tab \strong{Description} \tab \strong{Default}\cr
config   \tab character \tab config file path                       \tab pkgr.yml\cr
loglevel \tab character \tab level for logging                      \tab         \cr
library \tab character \tab library to install packag               \tab         \cr
thread   \tab integer   \tab number of threads to execute with      \tab
}

\if{html}{
\out{
</details>
}
}

\if{html}{
\out{
<br>
<details>
<summary> <span title='Click to Expand'> Global flags </span> </summary>
}
}

\tabular{ll}{
\strong{Flag} \tab \strong{Description}\cr
debug       \tab use debug mode\cr
preview     \tab preview action, but do not run command \cr
update      \tab update packages along with install
}

\if{html}{
\out{
</details>
}
}
}
\examples{

pkgr.add(c('dplyr','ggplot2'))

pkgr.add(c('dplyr','ggplot2'), config = 'pkgr.yml')

# here::here is built into pkgr.utils
pkgr.add(c('dplyr','ggplot2'), config = pkgr.here())

pkgr.add(c('dplyr','ggplot2'), config = 'pkgr.yml', flags = 'install')

pkgr.remove(c('dplyr','ggplot2'))

pkgr.install()

pkgr.install(config = 'pkgr.yml', library = .libPaths()[1])

pkgr.plan(config = 'pkgr.yml', library = .libPaths()[1])

}
