% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgr_commands.R
\name{pkgr.add}
\alias{pkgr.add}
\alias{pkgr.install}
\alias{pkgr.plan}
\title{FUNCTION_TITLE}
\usage{
pkgr.add(pkgs, ...)

pkgr.install(...)

pkgr.plan(...)
}
\arguments{
\item{pkgs}{PARAM_DESCRIPTION}

\item{\dots}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
