% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSetup_fns.R
\name{pkgSetup2pkgr}
\alias{pkgSetup2pkgr}
\title{FUNCTION_TITLE}
\usage{
pkgSetup2pkgr(pkgsetup = "pkgSetup.R", src_path = "pkg",
  lib_path = "lib", pkgr_tmpl = system.file("PKGSETUP.tmpl", package =
  "pkgr.utils"), out = "")
}
\arguments{
\item{pkgsetup}{PARAM_DESCRIPTION, Default: 'pkgSetup.R'}

\item{src_path}{PARAM_DESCRIPTION, Default: 'pkg'}

\item{lib_path}{PARAM_DESCRIPTION, Default: 'lib'}

\item{pkgr_tmpl}{PARAM_DESCRIPTION, Default: system.file("PKGSETUP.tmpl", package = "pkgr.utils")}

\item{out}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[yaml]{as.yaml}}
\code{\link[whisker]{whisker.render}}
}
